incorporated run_demo_workflow into niflow
