+++ niflow-simple-workflow/package/niflow/simple/workflow/__init__.py
@@ -1,0 +1,4 @@
+# Start with a minimal version
+# This should be bumped prior to a tag/release, or you may wish to consider using
+# Versioneer: https://github.com/warner/python-versioneer
+__version__ = '0.0.1'
