+++ niflow-simple-workflow/package/setup.cfg
@@ -1,0 +1,32 @@
+[metadata]
+name = niflow-simple-workflow
+version = attr: niflow.simple.workflow.__version__
+url = https://github.com/niflows/niflow-simple-workflow
+author = Yaroslav Halchenko
+author_email = debian@onerussian.com
+description = niflow-simple-workflow workflow package
+license = Apache License, 2.0
+classifiers =
+    Development Status :: 2 - Pre-Alpha
+    Environment :: Console
+    Intended Audience :: Developers
+    License :: OSI Approved :: Apache Software License
+    Operating System :: OS Independent
+    Programming Language :: Python :: 2
+    Topic :: Scientific/Engineering
+
+[options]
+packages = find:
+# Place any Python dependencies here
+# install_requires = 
+# Uncomment the following line if your package should distribute any non-Python files
+# include_package_data = True
+
+[options.entry_points]
+console_scripts =
+    niflow-simple-workflow = niflow.simple.workflow.cli:main
+
+## If your workflow requires data to run, uncomment and update the following section:
+# [options.package_data]
+# niflow-simple-workflow =
+#     niflow/simple/workflow/data/*
